Move NLI probing tasks
