+++ src/tasks/__init__.py
@@ -3,6 +3,7 @@ from . import tasks
 from . import edge_probing
 from . import lm
 from . import mt
+from . import nli_probing
 from . import reddit
 
 # REGISTRY needs to be available to modules within this package,
